Bug: Don't remove letters
##Descriptions

In the operation (activate function), when filtering to obtain the cards returned to the hand. It was done on the player (tp) and not on the opponent (1-tp). So it has changed.

##Context
In turn 2, storming mirror force bounce monster to hand and <Ruthless Drop Off> is used but nothing is banished
(See replay: https://cdn.discordapp.com/attachments/291708361752313866/411078324681441282/drop_off_bug.yrp)

##Screenshot:
![Alt Text](https://cdn.discordapp.com/attachments/291708361752313866/411099523159425024/2018-02-08_11-01-55.gif)