Various fixes
Fixed Millennium-Eyes Restrict: it was negating any equipped card
regardless if it was equipped by its own effect or not.
Legacy of the Duelist and Dragon Airship - Fandra:  If you destroy them
before the resolution of the effect that makes you skip the normal draw,
you should still be able to draw.