Fixed "Red Reboot"
Fixed an incorrect interaction with "Dinomight Knight, the True Dracofighter", where the player affect by "Red Reboot" would still be able to activate a "True Draco" or "True King" Continuous Trap Card from the Deck. The player should only be able to add it to hand in this scenario. Also includes a fix for a bug that your prevent the player from attempting to activate the effects of Trap Cards already face-up on the field.