Happy April Fools!
In our yearly tradition, we've added a bunch of joke cards to celebrate April Fools. This year we've added quite a few more. They all have "04/01" in their text, so search that in Deck Edit to find them. Then, use "Don't Check Deck" or "Allowed Cards:N/A" to use them in a Duel. Happy pranking!