Various fixes (#1872)
* "Delta Flyer" fix

Shouldn't be able to target a monster that doesn't have a Level (Xyz/Link).

* "Yosenju Wind Ritual" update

Removed leftover debug messages.

* "Fortune Lady Past" fix

Shouldn't be able to reduce the target's Level if it's already 1.

* "Mayosenju Hitot" fix

The Pendulum Effect always changed its own scale instead of the target's. Also, the Special Summon limitation applied after using it was incorrect.

* "Fortune Lady" updates

Updated the Special Summon handling for Fire, Light and Dark.

* "Fortune Lady Past" update

Fixed her effect for the 2nd player.