Various fixes/updates (#1979)
* "Salamangreat Fowl" fix

Its 1st effect should not trigger if the "Salamangreat" monster is summoned face-down.

* "The Weather Painter Cloud" fix

His 1st effect should not trigger if a "The Weather" card is sent to the GY from the opponent's field.

* "TGX3-DX2" fix

Should not be able to target "T.G." Spells/Traps.

* "Number 76: Gladiyell the Melody Lumaestra" fix

Her last effect should be a Quick Effect. Also, it should detach the Xyz Material as part of the effect and not as a cost.

* "Idaten the Conqueror Star" fix

Shouldn't be able to use his effect to make ATK 0 if the monster's ATK is already 0.

* "Flying Carp" fix

Fixed some typos with its second effect.

* "El Shaddoll Shekhinaga" fix

The negation effect should not destroy the monster if the activation of the effect was not negated successfully.

* "Phantom Fortress Enterblathnir" update

Added operation infos.

* "Phantom Fortress Enterblathnir" update

Small update.