Ritual procedure second update (#2282)
Reworked all the cards that perform a ritual summon, and merged them under the same utility for easier management. Fixed cards that can be treated as "Whole tribute" when the target monster has a reduced/increased level. Fixed ritual materials requirements for monsters with a changed level where the ritual card doesnt' specify the level to be matched