Master Rules 2020/Special Duel Modes update
Added visual indication that the rules/Duel Modes have been accepted by both players or declined by a player. The card's image will now flash on the screen if accepted (playing the "activation" animation). Also, in either case there will be a pop-up box saying whether both players accepted or not (that message will also be saved in the "Log" tab in the Duel).