Small fix "Adamacia Relieve"
Now you can tribute cards if there are an appropriate number of targets. (Previously, you could tribute more than you should be able to target if, and only if, the targets are just what are left over after tributing. Example: tribute 1 to target the only other card on the field.)